Remove raylib example because macOS fails to run coverage tests
Clang doesn't support code coverage reports via lcov, so we must
always use gcc for that.  However, building the graphics dependencies of
raylib on macOS fails when using gcc (it only works when using clang).

So for this project it is better to not use a raylib example rather than
biting the bullet and disabling code coverage reports.